+++ features/scripts/build-plugin-wsl.sh
@@ -8,15 +8,15 @@ UE_HOME="${UE_HOME:-C:\Program Files\Epic Games}"\\UE_$UE_VERSION
 
 UE_RUNUAT="$UE_HOME\Engine\Build\BatchFiles\RunUAT.bat"
 
-UE_TARGET_PLATFROM="Win32+Win64"
-
 if [[ "$UE_VERSION" == "5.1" ]]; then
-  UE_TARGET_PLATFROM="Win64"
+  TARGET_PLATFORMS="Win64"
+else
+  TARGET_PLATFORMS="Win32+Win64"
 fi
 
 SCRIPT_DIR=$(pwd)
 
-/mnt/c/windows/system32/cmd.exe /C "$UE_RUNUAT" BuildPlugin -Plugin="$(wslpath -w .)/plugins/Bugsnag/Bugsnag.uplugin" -Package="$(wslpath -w .)/Build/Plugin/Bugsnag" -TargetPlatforms="$UE_TARGET_PLATFROM"
+/mnt/c/windows/system32/cmd.exe /C "$UE_RUNUAT" BuildPlugin -Plugin="$(wslpath -w .)/plugins/Bugsnag/Bugsnag.uplugin" -Package="$(wslpath -w .)/Build/Plugin/Bugsnag" -TargetPlatforms="$TARGET_PLATFORMS"
 
 cd "$SCRIPT_DIR/Build/Plugin"